Use functions, ending with _utf8 under Windows
