+++ src/lib.rs
@@ -185,7 +185,8 @@ mod source;
 mod time_val;
 pub mod translate;
 pub mod types;
-pub mod utils;
+mod utils;
+pub use utils::*;
 pub mod value;
 pub mod variant;
 mod variant_type;
